Creado un component para reusarlo
