+++ src/types/common/icons.ts
@@ -1,0 +1,15 @@
+export const AccountIcons: string[] = [
+  "dollar-sign",
+  "file-invoice-dollar",
+  "euro-sign",
+  "hand-holding-usd",
+  "university",
+  "piggy-bank",
+  "money-bill",
+  "money-bill-alt",
+  "money-bill-wave",
+  "money-bill-wave-alt",
+  "money-check",
+  "money-check-alt",
+  "wallet",
+]
