+++ src/types/common/index.ts
@@ -1,2 +1,3 @@
 export * from "./constants"
 export * from "./types"
+export * from "./icons"
